Fixed Agent/Manager UTs in non Windows platforms.
