+++ src/unit_tests/os_execd/CMakeLists.txt
@@ -22,17 +22,15 @@ target_link_libraries(EXECD_O ${WAZUHLIB} ${WAZUHEXT} -lpthread)
 # Generate Execd tests
 if(${TARGET} STREQUAL "winagent")
     list(APPEND execd_names "test_win_execd")
-    list(APPEND execd_flags "-Wl,--wrap,_mdebug1 -Wl,--wrap,_merror \
+    list(APPEND execd_flags "-Wl,--wrap,_mtdebug1 -Wl,--wrap,_mterror \
                              -Wl,--wrap,read_exec_config -Wl,--wrap,get_command_by_name -Wl,--wrap,wpopenv -Wl,--wrap,wpclose -Wl,--wrap,fwrite \
                              -Wl,--wrap,fopen -Wl,--wrap,fclose -Wl,--wrap,fflush -Wl,--wrap,fgets -Wl,--wrap,fread -Wl,--wrap,fseek \
                              -Wl,--wrap,remove -Wl,--wrap,fgetpos -Wl,--wrap=fgetc")
 else()
     list(APPEND execd_names "test_execd")
-    list(APPEND execd_flags "-Wl,--wrap,_mdebug -Wl,--wrap,_mdebug1 -Wl,--wrap,_mdebug2 -Wl,--wrap,_merror \
-                             -Wl,--wrap,_mtdebug1 -Wl,--wrap,_mtdebug2 -Wl,--wrap,_mtwarn -Wl,--wrap,_mtinfo \
-                             -Wl,--wrap,select -Wl,--wrap,OS_RecvUnix -Wl,--wrap,CreateThreadJoinable -Wl,--wrap,read_exec_config \
-                             -Wl,--wrap,get_command_by_name -Wl,--wrap,wpopenv -Wl,--wrap,wpclose -Wl,--wrap,fwrite \
-                             -Wl,--wrap,fopen -Wl,--wrap,fclose -Wl,--wrap,fflush -Wl,--wrap,fgets -Wl,--wrap,fread -Wl,--wrap,fseek \
+    list(APPEND execd_flags "-Wl,--wrap,_mtdebug1 -Wl,--wrap,_mtdebug2 -Wl,--wrap,_mterror -Wl,--wrap,time -Wl,--wrap,select -Wl,--wrap,OS_RecvUnix \
+                             -Wl,--wrap,read_exec_config -Wl,--wrap,get_command_by_name -Wl,--wrap,wpopenv -Wl,--wrap,wpclose -Wl,--wrap,fwrite \
+                             -Wl,--wrap,fopen -Wl,--wrap,fclose -Wl,--wrap,fflush -Wl,--wrap,fgets -Wl,--wrap,fread -Wl,--wrap,fseek -Wl,--wrap,CreateThreadJoinable \
                              -Wl,--wrap,remove -Wl,--wrap,fgetpos -Wl,--wrap=fgetc")
 endif()
